+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * set `homepage` to `PluginURI` or `ThemeURI`, fixes [#791](https://github.com/afragen/github-updater/issues/791)
+* fixed Bitbucket release asset updates for proper containing folder structure
 
 #### 8.8.0 / 2019-05-15
 * switched from `pre_set_site_transient_update_{plugins|themes}` to `site_transient_update_{plugins|themes}`